casebreaker-backend/pyproject.toml
[project]
name = "casebreaker-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Imetomi",email = "imetstamas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "anthropic (>=0.46.0,<0.47.0)",
    "sseclient-py (>=1.8.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "casebreaker_backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
